Move throttle to its own file
