+++ lib/concurrent/edge/promises.rb
@@ -14,6 +14,35 @@ module Concurrent
   # {include:file:doc/promises-main.md}
   module Promises
 
+    # TODO (pitr-ch 23-Dec-2016): move out
+    # @!visibility private
+    module ReInclude
+      def included(base)
+        included_into << [:include, base]
+        super(base)
+      end
+
+      def extended(base)
+        included_into << [:extend, base]
+        super(base)
+      end
+
+      def include(*modules)
+        super(*modules)
+        modules.reverse.each do |module_being_included|
+          included_into.each do |method, mod|
+            mod.send method, module_being_included
+          end
+        end
+      end
+
+      private
+
+      def included_into
+        @included_into ||= []
+      end
+    end
+
     # @!macro [new] promises.param.default_executor
     #   @param [Executor, :io, :fast] default_executor Instance of an executor or a name of the
     #     global executor. Default executor propagates to chained futures unless overridden with
@@ -46,6 +75,7 @@ module Concurrent
     # Container of all {Future}, {Event} factory methods. They are never constructed directly with
     # new.
     module FactoryMethods
+      extend ReInclude
 
       # @!macro promises.shortcut.on
       # @return [ResolvableEvent]
@@ -1920,115 +1950,6 @@ module Concurrent
       end
 
       include ActorIntegration
-    end
-
-    # A tool manage concurrency level of future tasks.
-    # @example With futures
-    #   data     = (1..5).to_a
-    #   db       = data.reduce({}) { |h, v| h.update v => v.to_s }
-    #   max_two = Promises.throttle 2
-    #
-    #   futures = data.map do |data|
-    #     Promises.future(data) { |data|
-    #       # un-throttled, concurrency level equal data.size
-    #       data + 1
-    #     }.then_throttle(max_two, db) { |v, db|
-    #       # throttled, only 2 tasks executed at the same time
-    #       # e.g. limiting access to db
-    #       db[v]
-    #     }
-    #   end
-    #
-    #   futures.map(&:value!) # => [2, 3, 4, 5, nil]
-    #
-    # @example With Threads
-    #   max_two = Concurrent::Throttle.new 2
-    #   5.timse
-    class Throttle < Synchronization::Object
-      # TODO (pitr-ch 23-Dec-2016): move into different file
-      # TODO (pitr-ch 23-Dec-2016): move to Concurrent space
-      # TODO (pitr-ch 21-Dec-2016): consider using sized channel for implementation instead when available
-
-      safe_initialization!
-      private *attr_atomic(:can_run)
-
-      # New throttle.
-      # @param [Integer] limit
-      def initialize(limit)
-        super()
-        @Limit       = limit
-        self.can_run = limit
-        @Queue       = LockFreeQueue.new
-      end
-
-      # @return [Integer] The limit.
-      def limit
-        @Limit
-      end
-
-      def trigger
-        while true
-          current_can_run = can_run
-          if compare_and_set_can_run current_can_run, current_can_run - 1
-            if current_can_run > 0
-              return Promises.resolved_event
-            else
-              event = Promises.resolvable_event
-              @Queue.push event
-              return event
-            end
-          end
-        end
-      end
-
-      def release
-        while true
-          current_can_run = can_run
-          if compare_and_set_can_run current_can_run, current_can_run + 1
-            if current_can_run < 0
-              Thread.pass until (trigger = @Queue.pop)
-              trigger.resolve
-            end
-            return self
-          end
-        end
-      end
-
-      # @return [String] Short string representation.
-      def to_s
-        format '<#%s:0x%x limit:%s can_run:%d>', self.class, object_id << 1, @Limit, can_run
-      end
-
-      alias_method :inspect, :to_s
-
-      module PromisesIntegration
-        # TODO (pitr-ch 23-Dec-2016): apply similar pattern elsewhere
-
-        def throttled(&throttled_futures)
-          throttled_futures.call(trigger).on_resolution! { release }
-        end
-
-        def then_throttled(*args, &task)
-          trigger.then(*args, &task).on_resolution! { release }
-        end
-      end
-
-      include PromisesIntegration
-    end
-
-    class AbstractEventFuture < Synchronization::Object
-      module ThrottleIntegration
-        def throttled_by(throttle, &throttled_futures)
-          a_trigger = throttle.trigger & self
-          throttled_futures.call(a_trigger).on_resolution! { throttle.release }
-        end
-
-        def then_throttled_by(throttle, *args, &block)
-          throttled_by(throttle) { |trigger| trigger.then(*args, &block) }
-        end
-      end
-
-      include ThrottleIntegration
     end
 
     ### Experimental features follow
